cli/pyproject.toml
[project]
name = "helix-cli"
version = "0.1.0"
description = "CLI for Helix evolutionary prompt optimization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer[all]>=0.16.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.13.0",
    "jinja2>=3.1.6",
    "openai>=2.8.0",
    "tenacity>=9.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
]

[project.scripts]
helix = "helix_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["helix_cli"]

[tool.ruff]
line-length = 100
target-version = "py313"
